Rename msg.rs to proto.rs
